requirements.txt
selenium==4.10.0
behave==1.2.6
python-dotenv==1.0.0